Move SpecialGates to different file
