+++ src/qibo/gates/__init__.py
@@ -1,2 +1,3 @@
 from qibo.gates.gates import *
+from qibo.gates.special import *
 from qibo.gates.channels import *
